+++ src/integration/extradelight/java/plus/dragons/createcentralkitchen/integration/extradelight/mechanicalArm/EvaporatorArmInteractionPoint.java
@@ -1,0 +1,77 @@
+/*
+ * Copyright (C) 2025  DragonsPlus
+ * SPDX-License-Identifier: LGPL-3.0-or-later
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package plus.dragons.createcentralkitchen.integration.extradelight.mechanicalArm;
+
+import com.lance5057.extradelight.workstations.evaporator.EvaporatorBlockEntity;
+import com.simibubi.create.content.kinetics.mechanicalArm.ArmBlockEntity;
+import com.simibubi.create.content.kinetics.mechanicalArm.ArmInteractionPoint;
+import com.simibubi.create.content.kinetics.mechanicalArm.ArmInteractionPointType;
+import net.minecraft.core.BlockPos;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.phys.Vec3;
+import org.jetbrains.annotations.Nullable;
+
+public class EvaporatorArmInteractionPoint extends ArmInteractionPoint {
+    public EvaporatorArmInteractionPoint(ArmInteractionPointType type, Level level, BlockPos pos, BlockState state) {
+        super(type, level, pos, state);
+    }
+
+    @Override
+    public ItemStack extract(ArmBlockEntity armBlockEntity, int slot, int amount, boolean simulate) {
+        if (level.getBlockEntity(pos) instanceof EvaporatorBlockEntity evaporator) {
+            var inventory = evaporator.getItemHandler();
+            for (int i = 0; i < inventory.getSlots(); i++) {
+                var result = inventory.extractItem(i, amount, simulate);
+                if (!result.isEmpty())
+                    return result;
+            }
+        }
+        return ItemStack.EMPTY;
+    }
+
+    @Override
+    public ItemStack insert(ArmBlockEntity armBlockEntity, ItemStack stack, boolean simulate) {
+        return stack;
+    }
+
+    @Override
+    public int getSlotCount(ArmBlockEntity armBlockEntity) {
+        return 1;
+    }
+
+    @Override
+    protected Vec3 getInteractionPositionVector() {
+        return Vec3.upFromBottomCenterOf(pos, 13 / 16f);
+    }
+
+    public static class Type extends ArmInteractionPointType {
+        @Override
+        public boolean canCreatePoint(Level level, BlockPos pos, BlockState state) {
+            return level.getBlockEntity(pos) instanceof EvaporatorBlockEntity;
+        }
+
+        @Nullable
+        @Override
+        public ArmInteractionPoint createPoint(Level level, BlockPos pos, BlockState state) {
+            return new EvaporatorArmInteractionPoint(this, level, pos, state);
+        }
+    }
+}